refactor: move legacy apis to apis/cluster
