+++ apis/cluster/user/v1alpha1/zz_generated.managedlist.go
