+++ internal/controller/config/config.go
@@ -24,7 +24,7 @@ import (
 	"github.com/crossplane/crossplane-runtime/v2/pkg/resource"
 	ctrl "sigs.k8s.io/controller-runtime"
 
-	"github.com/statnett/provider-cloudian/apis/v1alpha1"
+	"github.com/statnett/provider-cloudian/apis/cluster/v1alpha1"
 )
 
 // Setup adds a controller that reconciles ProviderConfigs by accounting for
